% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_of_death_counts.R
\name{cause_of_death_counts}
\alias{cause_of_death_counts}
\title{A tidyUkBioBank function
Function output: counts data of all individuals who contained a matching string in the cause of death columns (M/F breakdown, etc.)}
\usage{
cause_of_death_counts(description, dataframe)
}
\arguments{
\item{description}{diagnosis of interest (string)}

\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tidyUkBioBank function
Function output: counts data of all individuals who contained a matching string in the cause of death columns (M/F breakdown, etc.)
}
\examples{
cause_of_death_counts()
}
\keyword{cause}
\keyword{counts}
\keyword{death}
\keyword{of}
