% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cause_of_death_table.R
\name{cause_of_death_table}
\alias{cause_of_death_table}
\title{A tidyUkBioBank function
Function output: a table with two columns: eid and cause of death description, which contains string matching the dscription variable given}
\usage{
get_cause_of_death_table(description, ukb_data)
}
\arguments{
\item{description}{diagnosis of interest (string)}

\item{ukb_data}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}

\item{disease_name}{string with disease name}
}
\description{
A tidyUkBioBank function
Function output: a table with two columns: eid and cause of death description, which contains string matching the dscription variable given
}
\examples{
get_cause_of_death_table()
}
\keyword{cause}
\keyword{death}
\keyword{get}
\keyword{of}
\keyword{table}
