% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_age.R
\name{cod_age}
\alias{cod_age}
\title{Make a cod_date table}
\usage{
cod_age(icd_list, disease_name, dataframe)
}
\arguments{
\item{icd_list}{list of icds}

\item{dataframe}{dataframe}
}
\description{
Make a cod_date table
}
\keyword{age}
\keyword{cod}
