% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_date.R
\name{cod_date}
\alias{cod_date}
\title{Make a cod_date table}
\usage{
cod_date(icd_list, disease_name, dataframe)
}
\arguments{
\item{icd_list}{list of icds}

\item{dataframe}{dataframe}
}
\description{
Make a cod_date table
}
\keyword{cod}
\keyword{date}
