% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses_ages.R
\name{diagnoses_ages}
\alias{diagnoses_ages}
\title{Make a diagnoses_ages table}
\usage{
diagnoses_ages(dataframe, ...)
}
\arguments{
\item{dataframe}{dataframe}
}
\description{
Make a diagnoses_ages table
}
\keyword{ages}
\keyword{diagnoses}
