% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses_dates.R
\name{diagnoses_dates}
\alias{diagnoses_dates}
\title{Make a diagnoses_dates table}
\usage{
diagnoses_dates(dataframe, ...)
}
\arguments{
\item{dataframe}{dataframe}
}
\description{
Make a diagnoses_dates table
}
\keyword{dates}
\keyword{diagnoses}
