% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses_table.R
\name{diagnoses_table}
\alias{diagnoses_table}
\title{A tidyUkBioBank function
Function output: a table with an eid columns, columns pertaining to specific icd/ self reported/ cause of death codes, and a column for all previously mentioned disease indicators (combined)
The columns with disease indicators contain 1's and 0's to represent whether the eid has the icd code, etc. of interest (1) or not (0)}
\usage{
diagnoses_table(icd_list, ukb_data, ...)
}
\arguments{
\item{icd_list}{list of icd codes}

\item{ukb_data}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}

\item{self_reported}{optional. if the user requests self reported diagnoses, then type self_reported = sr_list, where sr_list contains the self_reported codes of interest}

\item{cause_of_death}{optional. if the user requests cause_of_death diagnoses, then type cause_of_death = cod_list, where cod_list contains the cause of death of interest}
}
\description{
A tidyUkBioBank function
Function output: a table with an eid columns, columns pertaining to specific icd/ self reported/ cause of death codes, and a column for all previously mentioned disease indicators (combined)
The columns with disease indicators contain 1's and 0's to represent whether the eid has the icd code, etc. of interest (1) or not (0)
}
\examples{
diagnoses_table()
}
\keyword{diagnoses}
\keyword{table}
