% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cod.R
\name{dx_cod}
\alias{dx_cod}
\title{Make a dx_cod table}
\usage{
dx_cod(icd_list, disease_name, ukb_data)
}
\arguments{
\item{icd_list}{list of icds}

\item{ukb_data}{dataframe}
}
\description{
Make a dx_cod table
}
\keyword{cod}
\keyword{dx}
