% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_demographics.R
\name{dx_demographics}
\alias{dx_demographics}
\title{A tidyUkBioBank function
Function output: dataframe containing counts data for icds of interest, self reported diseases (via codes or strings), and cause of death information (strings)}
\usage{
dx_demographics(icd_list, dataframe, ...)
}
\arguments{
\item{icd_list}{list of icds of interest}

\item{ukb_data}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tidyUkBioBank function
Function output: dataframe containing counts data for icds of interest, self reported diseases (via codes or strings), and cause of death information (strings)
}
\examples{
dx_demographics()
}
\keyword{demographics}
\keyword{dx}
