% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_sr.R
\name{dx_sr}
\alias{dx_sr}
\title{Make a dx_sr table}
\usage{
dx_sr(diagnosis, disease_name, ukb_data, cancer)
}
\arguments{
\item{diagnosis}{list of icds}

\item{disease_name}{disease_name}

\item{ukb_data}{ukb_data}

\item{cancer}{cancer}
}
\description{
Make a dx_sr table
}
\keyword{dx}
\keyword{sr}
