% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chromosome_list.R
\name{get_chromosome_list}
\alias{get_chromosome_list}
\title{A tidyUkBioBank function
Function output: list with all of the chromosomes contianing full pgen/psam/pvar sets in the directory}
\usage{
get_chromosome_list(pattern, directory)
}
\arguments{
\item{pattern}{the pattern by which the chromosome of the file is indicated (ex: if the files have the pattern "c1.pgen", "c2.psam", etc. then the pattern is "c")}

\item{directory}{the directory containing the pgens, psams, and pvars}
}
\description{
A tidyUkBioBank function
Function output: list with all of the chromosomes contianing full pgen/psam/pvar sets in the directory
}
\examples{
get_chromosome_list()
}
\keyword{chromosome}
\keyword{get}
\keyword{list}
