% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chromosome_variants.R
\name{get_chromosome_variants}
\alias{get_chromosome_variants}
\title{A tidyUkBioBank function
Function output: dataframe containing all of the variants on the specified chromosome}
\usage{
get_chromosome_variants(chromosome, directory)
}
\arguments{
\item{chromosome}{chromosome on which you are requesting variants}

\item{directory}{directory containing the psams, pgens, and pvars}
}
\description{
A tidyUkBioBank function
Function output: dataframe containing all of the variants on the specified chromosome
}
\examples{
get_chromosome_variants()
}
\keyword{chromosome}
\keyword{get}
\keyword{variants}
