% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_directory_variants.R
\name{get_directory_variants}
\alias{get_directory_variants}
\title{A tidyUkBioBank function
Function output: dataframe containing all of the variants on all of the chromosomes in the directory}
\usage{
get_directory_variants(pattern, directory)
}
\arguments{
\item{pattern}{pattern by which the chromsoomes are labeled ("c19" -> "c", "chr19" -> "chr")}

\item{directory}{directory containing the psams, pgens, and pvars}
}
\description{
A tidyUkBioBank function
Function output: dataframe containing all of the variants on all of the chromosomes in the directory
}
\examples{
get_directory_variants()
}
\keyword{directory}
\keyword{get}
\keyword{variants}
