% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_date.R
\name{get_first_date}
\alias{get_first_date}
\title{Make a get_first_date table}
\usage{
get_first_date(diagnosis_date_rows, diagnosis_date_group)
}
\description{
Make a get_first_date table
}
\keyword{get_first_date}
