% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotypes.R
\name{get_genotypes}
\alias{get_genotypes}
\title{A tidyUkBioBank function
Function output: dataframe containing eid column and columns with the genotype information for the variants of interest}
\usage{
get_genotypes(variants, variant_df, directory)
}
\arguments{
\item{directory}{directory containing the psams, pgens, and pvars}
}
\description{
A tidyUkBioBank function
Function output: dataframe containing eid column and columns with the genotype information for the variants of interest
}
\examples{
get_genotypes()
}
\keyword{genotypes}
\keyword{get}
