% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pgen_list.R
\name{get_pgen_list}
\alias{get_pgen_list}
\title{A tidyUkBioBank function
Function output: list of pgens}
\usage{
get_pgen_list(chr_list, directory)
}
\arguments{
\item{chr_list}{list of all of the chromosomes you are interested in recieving the variants of}

\item{directory}{the directory containing the pgens, psams, and pvars}
}
\description{
A tidyUkBioBank function
Function output: list of pgens
}
\examples{
get_pgen_list()
}
\keyword{get}
\keyword{list}
\keyword{pgen}
