% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_self_reported_eids.R
\name{get_self_reported_eids}
\alias{get_self_reported_eids}
\title{A tidyUkBioBank function
Function output: list of eids of individuals who self reported the diagnosis/ code of interest}
\usage{
get_self_reported_eids(diagnosis, ukb_data)
}
\arguments{
\item{diagnosis}{diagnosis of interest (string or code)}

\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tidyUkBioBank function
Function output: list of eids of individuals who self reported the diagnosis/ code of interest
}
\examples{
get_self_reported_eids()
}
\keyword{eids}
\keyword{get}
\keyword{reported}
\keyword{self}
