% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_self_reported_info.R
\name{get_self_reported_info}
\alias{get_self_reported_info}
\title{A tidyUkBioBank function
Function output: row of the coding6 dataframe, which connects self reported diagnoses to their codes}
\usage{
get_self_reported_info(disease_list)
}
\arguments{
\item{element}{a number code or string describing the self reported disease of interest}
}
\description{
A tidyUkBioBank function
Function output: row of the coding6 dataframe, which connects self reported diagnoses to their codes
}
\examples{
get_self_reported_info()
}
\keyword{info}
\keyword{reported}
\keyword{self}
