% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_self_reported_table.R
\name{get_self_reported_table}
\alias{get_self_reported_table}
\title{A tidyUkBioBank function
Function output: table with the code/ string provided and information about the self_reported code selected
Outputs a print statment with information about the self reported code selected}
\usage{
get_self_reported_table(diagnosis, ukb_data)
}
\arguments{
\item{diagnosis}{a number code or string describing the self reported disease of interest}

\item{ukb_data}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tidyUkBioBank function
Function output: table with the code/ string provided and information about the self_reported code selected
Outputs a print statment with information about the self reported code selected
}
\examples{
get_self_reported_table()
}
\keyword{get}
\keyword{reported}
\keyword{self}
\keyword{table}
