% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats_by_sex.R
\name{get_stats_by_sex}
\alias{get_stats_by_sex}
\title{A tidyUkBioBank function
Function output: counts, mean, and median values for the genetic sex of your choice (Male or Female)}
\usage{
get_stats_by_sex(sex, dataframe)
}
\arguments{
\item{sex}{sex of interest, can be Male or Female}

\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tidyUkBioBank function
Function output: counts, mean, and median values for the genetic sex of your choice (Male or Female)
}
\examples{
get_stats_by_sex()
}
\keyword{by}
\keyword{sex}
\keyword{stats}
