% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variants.R
\name{get_variants}
\alias{get_variants}
\title{A tidyUkBioBank function
Function output: dataframe contianing all of the variants on the chromosomes requested}
\usage{
get_variants(chr_list, directory)
}
\arguments{
\item{chr_list}{list of all of the chromosomes you are interested in recieving the variants of}

\item{directory}{the directory containing the pgens, psams, and pvars}
}
\description{
A tidyUkBioBank function
Function output: dataframe contianing all of the variants on the chromosomes requested
}
\examples{
get_variants()
}
\keyword{get}
\keyword{variants}
