% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pgen.R
\name{make_pgen}
\alias{make_pgen}
\title{A tools4ukbb function
Function output: a pgen object from the chromosome of your choosing}
\usage{
make_pgen(chromosome, directory)
}
\arguments{
\item{chromosome}{which chromosome you are selecting and how it is labeled (eg: for chromosome 2 of sample3_chr2.pgen, chromsome = "chr2" OR for s1_c2.pgen, chromosome = "c2")}

\item{directory}{the directory with your pgen files}
}
\description{
A tools4ukbb function
Function output: a pgen object from the chromosome of your choosing
}
\examples{
make_pgen()
}
\keyword{pgen}
