% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_psam.R
\name{make_psam}
\alias{make_psam}
\title{A tools4ukbb function
Function output: a tsv with psam details from the chromosome of your choosing}
\usage{
make_psam(chromosome, directory)
}
\arguments{
\item{chromosome}{which chromosome you are selecting and how it is labeled (eg: for chromosome 2 of sample3_chr2.psam, chromsome = "chr2" OR for s1_c2.psam, chromosome = "c2")}

\item{directory}{the directory with your psam files}
}
\description{
A tools4ukbb function
Function output: a tsv with psam details from the chromosome of your choosing
}
\examples{
make_psam()
}
\keyword{psam}
