% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_get_SR_table_input.R
\name{parse_get_SR_table_input}
\alias{parse_get_SR_table_input}
\title{A tidyUkBioBank function
Function output: internal function, code indicated by either provided code or string
Outputs a print statment with information about the self reported code selected}
\usage{
parse_get_SR_table_input(disease)
}
\arguments{
\item{disease}{a number code or string describing the self reported disease of interest}
}
\description{
A tidyUkBioBank function
Function output: internal function, code indicated by either provided code or string
Outputs a print statment with information about the self reported code selected
}
\examples{
parse_get_SR_table_input()
}
\keyword{info}
\keyword{parse}
\keyword{reported}
\keyword{self}
