% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_coding.R
\name{search_coding}
\alias{search_coding}
\title{A tidyUkBioBank function
Function output: row of the coding6 dataframe, which connects self reported diagnoses to their codes}
\usage{
search_coding(element)
}
\arguments{
\item{element}{a number code describing the self reported disease of interest}
}
\description{
A tidyUkBioBank function
Function output: row of the coding6 dataframe, which connects self reported diagnoses to their codes
}
\examples{
search_coding()
}
\keyword{coding}
\keyword{search}
