% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_icd_codes.R
\name{search_icd_codes}
\alias{search_icd_codes}
\title{A tidyUkBioBank function
Function output: internal function, a list of individuals whose data contains the icd codes in the icd_list variable}
\usage{
search_icd_codes(columns, icd_list, ukb_data)
}
\arguments{
\item{columns}{columns to be searched}

\item{icd_list}{list of icd codes}

\item{ukb_data}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tidyUkBioBank function
Function output: internal function, a list of individuals whose data contains the icd codes in the icd_list variable
}
\examples{
search_icd_codes()
}
\keyword{codes}
\keyword{icd}
\keyword{search}
