% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_meaning.R
\name{search_meaning}
\alias{search_meaning}
\title{A tidyUkBioBank function
Function output: row of the coding6 dataframe, which connects self reported diagnoses to their codes}
\usage{
search_meaning(element)
}
\arguments{
\item{element}{a string describing the self reported disease of interest (ex: "Lupus", "Dementia")}
}
\description{
A tidyUkBioBank function
Function output: row of the coding6 dataframe, which connects self reported diagnoses to their codes
}
\examples{
search_meaning()
}
\keyword{meaning}
\keyword{search}
