% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_age_stats.R
\name{sex_age_stats}
\alias{sex_age_stats}
\title{A tidyUkBioBank function
Function output: dataframe containing counts data and age of diagnosis inforamtion for icd codes of interest}
\usage{
sex_age_stats(list_icd, dataframe)
}
\arguments{
\item{list_icd}{list of icds of interest}

\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tidyUkBioBank function
Function output: dataframe containing counts data and age of diagnosis inforamtion for icd codes of interest
}
\examples{
sex_age_stats()
}
\keyword{sex_age_stats}
