% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_age.R
\name{sr_age}
\alias{sr_age}
\title{Make a sr_age table}
\usage{
sr_age(diagnosis, disease_name, ukb_data, cancer)
}
\arguments{
\item{diagnosis}{list of icds}

\item{disease_name}{disease_name}

\item{ukb_data}{ukb_data}

\item{cancer}{cancer}
}
\description{
Make a sr_age table
}
\keyword{age}
\keyword{sr}
