% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_date.R
\name{sr_date}
\alias{sr_date}
\title{Make a sr_date table}
\usage{
sr_date(diagnosis, disease_name, dataframe, cancer)
}
\arguments{
\item{diagnosis}{list of icds}

\item{disease_name}{disease_name}

\item{dataframe}{ukb_data}

\item{cancer}{cancer}
}
\description{
Make a sr_date table
}
\keyword{date}
\keyword{sr}
