% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_info.R
\name{ukb_info}
\alias{ukb_info}
\title{A tidyUkBioBank function
Function output: dataframe containing eids, date of first diagnosis, and age at first recorded daignosis}
\usage{
ukb_info(icd_list, dataframe, disease_name)
}
\arguments{
\item{icd_list}{list of icds of interest}

\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}

\item{disease_name}{name of the disease of interest (under which all of the icd codes in icd_list fall)}
}
\description{
A tidyUkBioBank function
Function output: dataframe containing eids, date of first diagnosis, and age at first recorded daignosis
}
\examples{
ukb_info()
}
\keyword{ukb_info}
